no voting for sharing transactions
